+++ dasp_window/src/hann/mod.rs
@@ -6,16 +6,16 @@ mod ops;
 
 /// A type of window function, also known as the "raised cosine window".
 ///
-/// [Wiki entry](https://en.wikipedia.org/wiki/Window_function#Hann_.28Hanning.29_window).
+/// [Wiki entry](https://en.wikipedia.org/wiki/Hann_function).
 ///
 /// ### Required Features
 ///
-/// - When using `dasp_window`, this item requires the **hanning** feature to be enabled.
-/// - When using `dasp`, this item requires the **window-hanning** feature to be enabled.
+/// - When using `dasp_window`, this item requires the **hann** feature to be enabled.
+/// - When using `dasp`, this item requires the **window-hann** feature to be enabled.
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct Hanning;
+pub struct Hann;
 
-impl<S> Window<S> for Hanning
+impl<S> Window<S> for Hann
 where
     S: Sample,
 {